requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.10.0
langchain==0.0.335
google-generativeai==0.8.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic-settings==2.0.3
pytest==7.4.3
httpx==0.25.2